rtmp-services: Fix comparison mismatch warning

+++ plugins/rtmp-services/rtmp-common.c
@@ -778,8 +778,8 @@ static int get_bitrate_matrix_max(json_t *array)
 		if (c != 2)
 			continue;
 
-		if (ovi.output_width == cx && ovi.output_height == cy &&
-		    cur_fps <= fps)
+		if ((int)ovi.output_width == cx &&
+		    (int)ovi.output_height == cy && cur_fps <= fps)
 			return bitrate;
 	}
